renames index.tsx to index.ts
