+++ src/signals/incident/components/form/ContainerSelect/ContainerSelect.tsx
@@ -1,7 +1,7 @@
 import type { FunctionComponent } from 'react';
 import React, { useCallback, useState } from 'react';
 import { ContainerSelectProvider } from './context';
-import Intro from './Intro/Intro';
+import Intro from './Intro';
 import Selector from './Selector';
 import Summary from './Summary';
 import type { ClickEventHandler, Item, Meta } from './types';
