+++ src/signals/incident/components/form/ContainerSelect/Selector/LegendPanel/index.ts
