+++ src/signals/incident/components/form/ContainerSelect/Selector/WfsLayer/ContainerLayer/index.ts
@@ -1,0 +1,3 @@
+import ContainerLayer from './ContainerLayer';
+
+export default ContainerLayer;
