+++ src/signals/incident/components/form/ContainerSelect/Summary/index.ts
